+++ abstractions/go/authentication/anonymous_authentication_provider.go
@@ -1,0 +1,10 @@
+package authentication
+
+import abs "github.com/microsoft/kiota/abstractions/go"
+
+type AnonymousAuthenticationProvider struct {
+}
+
+func (provider *AnonymousAuthenticationProvider) Authenticate(request abs.RequestInfo) error {
+	return nil
+}