+++ abstractions/ruby/microsoft_kiota_abstractions/lib/microsoft_kiota_abstractions/authentication/anonymous_authentication_provider.rb
@@ -1,0 +1,10 @@
+require 'concurrent'
+
+module MicrosoftKiotaAbstractions
+  class AnonymousAuthenticationProvider
+    include MicrosoftKiotaAbstractions::AuthenticationProvider
+    include Concurrent::Async
+    def authenticate_request(request)
+    end
+  end
+end