+++ abstractions/typescript/src/authentication/index.ts
@@ -1,0 +1,3 @@
+export * from './authenticationProvider';
+export * from './anonymousAuthenticationProvider';
+export * from './baseBearerTokenAuthenticationProvider';