+++ abstractions/typescript/src/index.ts
@@ -3,7 +3,7 @@ export * from "./httpCore";
 export * from "./httpMethod";
 export * from "./requestInfo";
 export * from "./responseHandler";
-export * from "./authenticationProvider";
+export * from "./authentication";
 export * from "./nativeResponseHandler";
 export * from "./nativeResponseWrapper";
 export * from './serialization';